pyproject.toml
[project]
name = "langchain-training"
version = "0.1.0"
description = "Experimentos com LangChain"
authors = [
	{ name = "Samuel Machado" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"langchain",
	"transformers",
	"python-dotenv",
	"langchain-ollama>=1.0.1",
	"langchain-community>=0.4.1",
	"langchain-groq>=1.1.2",
	"beautifulsoup4>=4.14.3",
	"langchain-chroma>=1.1.0",
	"wikipedia>=1.4.0",
	"faiss-cpu>=1.13.2",
	"fastapi>=0.135.1",
	"uvicorn>=0.40.0",
	"youtube-transcript-api>=1.2.4",
	"pytube>=15.0.0",
	"streamlit>=1.55.0",
	"langchain-huggingface>=1.2.1",
	"pypdf>=6.7.5",
]

[project.optional-dependencies]
dev = []


